+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-Forum-client/src/test/java/com/graphql_java_generator/samples/forum/SpringTestConfig.java
@@ -6,10 +6,12 @@ import org.springframework.context.annotation.ComponentScan.Filter;
 import org.springframework.context.annotation.FilterType;
 
 import com.graphql_java_generator.client.GraphQLConfiguration;
+import com.graphql_java_generator.client.graphqlrepository.EnableGraphQLRepositories;
 
 @TestConfiguration
 @ComponentScan(basePackageClasses = { SpringTestConfig.class, GraphQLConfiguration.class }, excludeFilters = {
 		@Filter(type = FilterType.REGEX, pattern = "graphql\\..*") })
+@EnableGraphQLRepositories({ "com.graphql_java_generator.samples.forum.client.graphql" })
 public class SpringTestConfig {
 
 }