+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-allGraphQLCases-client/src/main/java/org/allGraphQLCases/graphqlrepository/non_spring/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Demonstrate the use of GraphQL Repositories in a non-Spring application
+ */
+package org.allGraphQLCases.graphqlrepository.non_spring;